Gangplank: Hop Pod (aka remote cluster mode)
During a demo, an engineer asked "when will I be able to run Gangplank
locally and have a cluster run it?" This commit provides for remote
execution in a remote cluster.

The purpouse of remote cluster support is two fold:
- Prow is too restrictive to run Gangaplnk in Pod mode. Simply put, the
  ci-builder lacks permissions to discover its own IP. Prow's ci-builder
  does have permissions to create pods in the coreos namespace.
- Developers who want to run a job in another cluster.

To accomplish this, Gangplank uses a "hop" pod in the remote cluster:
- local Gangplank starts a pod remotely
- the remote pod creates other pods as needed
- work is returned to the remote hop pod
- remote logs are streamed to the local pod

All remote results remains on the hop pod except for the logs.

Example:
  gangplank pod --remoteKubeConfig kubec.config <ARGS>

Signed-off-by: Ben Howard <ben.howard@redhat.com>