kola/tests/misc/network: Add test to check dhcp propagation over bond via veth
Add a new ci test, that will check dhcp propagation on dhcp bond using
beth pairs.
This test creates 2 veth pairs, with one end set on another namespace
"container".
the host side of the veth ends get a specific MAC-address
and the container side veth ends get a specific IP-Address.
the two host-side veth end are then connected to a bond.

In a later commit, we will connected a dnsmasq container to the
container network namespace, which will cause the host side bond to get
an ip.

Signed-off-by: Ram Lavi <ralavi@redhat.com>