Mantle/Kola: Fix race condition in QMP Socket commands
After turning on the race detector to fix another race condition in the
JSON reporter code, a race condition in the qmp socket handling was
discovered.

This fixes a race condition where different test go-routines can run
commands on the same qmp.socket by using a mutex. The prior code would
open a connection to the socket and hold it until completion, preventing
another go-routine from using it.

Signed-off-by: Ben Howard <ben.howard@redhat.com>